requirements.txt
pandas==2.2.2
numpy==1.26.4
polars==0.20.0
pyarrow==14.0.0
google-cloud-bigquery==3.16.0
google-cloud-storage==2.12.0
db-dtypes==1.2.0
jupyter==1.0.0
matplotlib==3.8.0
scikit-learn==1.4.2
lightgbm==4.3.0
joblib==1.4.0
tqdm==4.66.1
shap==0.44.1
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.33.0
pydantic==2.5.0
